AS3.x対応と各ライブラリ最新化
* Android Studio 3.x対応
* 各ライブラリを最新化
* ApplicationIDとPackage名変更
