+++ app/src/debug/AndroidManifest.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.kgyshuhei.kotlin_okhttp_retrofit">
+    package="com.shuheikgy.android_mvvm_architecture">
 
     <application
         android:name=".application.DebugBaseApplication"
