+++ app/src/debug/java/com/shuheikgy/android_mvvm_architecture/application/DebugBaseApplication.kt
@@ -1,4 +1,4 @@
-package com.kgyshuhei.kotlin_okhttp_retrofit.application
+package com.shuheikgy.android_mvvm_architecture.application
 
 import android.app.Application
 
@@ -6,7 +6,7 @@ import android.app.Application
  * Created by shuhei_k on 2017/07/08.
  */
 
-class BaseApplication : Application() {
+class DebugBaseApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
