+++ app/src/main/java/com/shuheikgy/android_mvvm_architecture/application/BaseApplication.kt
@@ -1,4 +1,4 @@
-package com.kgyshuhei.kotlin_okhttp_retrofit.application
+package com.shuheikgy.android_mvvm_architecture.application
 
 import android.app.Application
 
@@ -6,7 +6,7 @@ import android.app.Application
  * Created by shuhei_k on 2017/07/08.
  */
 
-class DebugBaseApplication : Application() {
+class BaseApplication : Application() {
 
     override fun onCreate() {
         super.onCreate()
