+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">kotlin-okhttp-retrofit</string>
+    <string name="app_name">Android MVVM Architecture Example</string>
 </resources>
